+++ hs-bindgen/src-internal/HsBindgen/Frontend/Pass/PrepareReparse/Update.hs
@@ -1,0 +1,266 @@
+-- | Updater for 'ReparseInfo' after preprocessing
+--
+-- This module is intended to be imported unqualified. It is also intended to
+-- only be imported from within the "HsBindgen.Frontend.Pass.PrepareReparse"
+-- module hierarchy.
+--
+-- > import HsBindgen.Frontend.Pass.PrepareReparse.Update
+--
+module HsBindgen.Frontend.Pass.PrepareReparse.Update (
+    update
+  ) where
+
+import Prelude hiding (lex, print)
+
+import Control.Monad.Reader (MonadReader (ask), ReaderT (..))
+import Control.Monad.State (MonadState, State, modify, runState)
+import Data.Kind (Type)
+import Data.Map.Lazy qualified as Map
+
+import Clang.HighLevel.Types qualified as Clang
+
+import HsBindgen.Errors (panicPure)
+import HsBindgen.Frontend.Analysis.DeclIndex (DeclIndex)
+import HsBindgen.Frontend.Analysis.DeclIndex qualified as DeclIndex
+import HsBindgen.Frontend.AST.Coerce (CoercePass (coercePass))
+import HsBindgen.Frontend.AST.Decl qualified as C
+import HsBindgen.Frontend.AST.TranslationUnit qualified as C
+import HsBindgen.Frontend.DeclMeta (DeclMeta (declIndex))
+import HsBindgen.Frontend.Naming (DeclId)
+import HsBindgen.Frontend.Pass.Parse.IsPass (ReparseInfo (ReparseNeeded, ReparseNotNeeded),
+                                             Tokens)
+import HsBindgen.Frontend.Pass.Parse.Msg (DelayedParseMsg (ParseMacroPrepareReparseFailed))
+import HsBindgen.Frontend.Pass.PrepareReparse.AST (Decl (..), Tag (..),
+                                                   TagType (Function))
+import HsBindgen.Frontend.Pass.PrepareReparse.Flatten (flattenDefault,
+                                                       flattenFunction)
+import HsBindgen.Frontend.Pass.PrepareReparse.IsPass (FlatTokens (..),
+                                                      PrepareReparse)
+import HsBindgen.Frontend.Pass.PrepareReparse.Simplifier (fieldTag, functionTag,
+                                                          typedefTag,
+                                                          variableTag)
+import HsBindgen.Frontend.Pass.TypecheckMacros.IsPass (CheckedMacro,
+                                                       TypecheckMacros)
+import HsBindgen.Imports (Map)
+
+{-------------------------------------------------------------------------------
+  Top-level
+-------------------------------------------------------------------------------}
+
+update ::
+     Maybe (Map Tag Decl)
+  -> C.TranslationUnit TypecheckMacros
+  -> C.TranslationUnit PrepareReparse
+update mapping unit = unit' {
+      C.meta = unit'.meta {
+          declIndex = declIndex'
+        }
+    }
+  where
+    (unit', msgs) = runM env $ updateIt () unit
+
+    declIndex' ::  DeclIndex
+    declIndex' =
+      -- We use @foldr@ here to establish the original order of messages
+      foldr
+        DeclIndex.registerDelayedParseMsg
+        unit.meta.declIndex
+        msgs
+
+    env = Env {
+        map = mapping
+      }
+
+{-------------------------------------------------------------------------------
+  Update: class
+-------------------------------------------------------------------------------}
+
+class Update a where
+  type Ctx a :: Type
+  type Ctx a = C.DeclInfo TypecheckMacros
+  updateIt :: Ctx a -> a TypecheckMacros -> M (a PrepareReparse)
+
+{-------------------------------------------------------------------------------
+  Update: monad
+-------------------------------------------------------------------------------}
+
+runM :: Env -> M a -> (a, [(DeclId, DelayedParseMsg)])
+runM m (M k) = fmap (.messages) $ runState (runReaderT k m) (St [])
+
+newtype M a = M (ReaderT Env (State St) a)
+  deriving newtype (Functor, Applicative, Monad)
+
+deriving newtype instance MonadReader Env M
+deriving newtype instance MonadState St M
+
+newtype Env = Env {
+    -- 'Nothing' means we could not expand macro invocations (for any of a
+    -- variety of reasons). We default to just flattening tokens at this
+    -- point.
+    map :: Maybe (Map Tag Decl)
+  }
+
+newtype St = St {
+    messages :: [(DeclId, DelayedParseMsg)]
+  }
+
+{-------------------------------------------------------------------------------
+  Update: instances
+-------------------------------------------------------------------------------}
+
+instance Update C.TranslationUnit where
+  type Ctx C.TranslationUnit = ()
+  updateIt _ unit = do
+      decls' <- mapM (updateIt ()) unit.decls
+      pure C.TranslationUnit {
+          decls = decls'
+        , includeGraph = unit.includeGraph
+        , meta = unit.meta
+        }
+
+instance Update C.Decl where
+  type Ctx C.Decl = ()
+  updateIt _ decl = do
+      kind' <- updateIt decl.info decl.kind
+      pure C.Decl {
+          info = coercePass decl.info
+        , kind = kind'
+        , ann = decl.ann
+        }
+
+instance Update C.DeclKind where
+  updateIt info declKind = case declKind of
+      C.DeclStruct struct      -> C.DeclStruct           <$> recurse struct
+      C.DeclUnion union        -> C.DeclUnion            <$> recurse union
+      C.DeclTypedef typedef    -> C.DeclTypedef          <$> recurse typedef
+      C.DeclEnum enum          -> C.DeclEnum             <$> recurse enum
+      C.DeclAnonEnumConstant c -> C.DeclAnonEnumConstant <$> recurse c
+      C.DeclOpaque             -> pure C.DeclOpaque
+      C.DeclMacro macro        -> C.DeclMacro            <$> recurse macro
+      C.DeclFunction function  -> C.DeclFunction         <$> recurse function
+      C.DeclGlobal global      -> C.DeclGlobal           <$> recurse global
+    where
+      recurse :: forall a.
+           (Update a, Ctx a ~ C.DeclInfo TypecheckMacros)
+        => a TypecheckMacros
+        -> M (a PrepareReparse)
+      recurse = updateIt info
+
+instance Update C.Struct where
+  updateIt info struct = do
+      fields' <- mapM (updateIt info) struct.fields
+      flam' <- mapM (updateIt info) struct.flam
+      pure C.Struct {
+          sizeof = struct.sizeof
+        , alignment = struct.alignment
+        , fields = fields'
+        , flam = flam'
+        , ann = struct.ann
+        }
+
+instance Update C.StructField where
+  updateIt info field = do
+      ann' <- updateReparseInfo info (fieldTag info field.info) field.ann
+      pure C.StructField {
+          info = coercePass field.info
+        , typ = coercePass field.typ
+        , offset = field.offset
+        , width = field.width
+        , ann = ann'
+        }
+
+instance Update C.Union where
+  updateIt info union = do
+      fields' <- mapM (updateIt info) union.fields
+      pure C.Union {
+          sizeof = union.sizeof
+        , alignment = union.alignment
+        , fields = fields'
+        , ann = union.ann
+        }
+
+instance Update C.UnionField where
+  updateIt info field = do
+      ann' <- updateReparseInfo info (fieldTag info field.info) field.ann
+      pure C.UnionField {
+          info = coercePass field.info
+        , typ = coercePass field.typ
+        , ann = ann'
+        }
+
+instance Update C.Typedef where
+  updateIt info typedef = do
+      ann' <- updateReparseInfo info (typedefTag info) typedef.ann
+      pure C.Typedef {
+          typ = coercePass typedef.typ
+        , ann = ann'
+        }
+
+instance Update C.Enum where
+  updateIt _ enum = pure $ coercePass enum
+
+instance Update C.AnonEnumConstant where
+  updateIt _ constant = pure $ coercePass constant
+
+instance Update CheckedMacro where
+  updateIt _info macro = pure $ coercePass macro
+
+instance Update C.Function where
+  updateIt info function = do
+      ann' <- updateReparseInfo info (functionTag info) function.ann
+      pure C.Function {
+          args = map coercePass function.args
+        , res = coercePass function.res
+        , attrs = function.attrs
+        , ann = ann'
+        }
+
+instance Update C.Global where
+  updateIt info global = do
+      ann' <- updateReparseInfo info (variableTag info) global.ann
+      pure C.Global {
+          typ = coercePass global.typ
+        , ann = ann'
+        }
+
+updateReparseInfo ::
+     C.DeclInfo TypecheckMacros
+  -> Tag
+  -> ReparseInfo Tokens
+  -> M (ReparseInfo FlatTokens)
+updateReparseInfo info tag@(Tag typ _) reparseInfo = do
+    env <- ask
+    case reparseInfo of
+      ReparseNotNeeded -> pure ReparseNotNeeded
+      ReparseNeeded tokens usedMacros -> do
+        let fallback = case typ of
+              Function -> flattenFunction tokens
+              _        -> flattenDefault tokens
+        flatten <-
+          case env.map of
+            -- If this is 'Nothing', we have already traced a reason why (see
+            -- 'PrepareReparseMsg').
+            Nothing -> pure fallback
+            Just mapping -> case Map.lookup tag mapping of
+              Nothing -> do
+                addMessage info.id ParseMacroPrepareReparseFailed
+                pure fallback
+              Just (Decl dec) -> pure dec
+        let flatTokens = FlatTokens {
+                flatten = flatten
+              , locStart = getLocation tokens
+              }
+        pure $ ReparseNeeded flatTokens usedMacros
+
+{-------------------------------------------------------------------------------
+  Internal auxiliary
+-------------------------------------------------------------------------------}
+
+getLocation :: [Clang.Token a] -> Clang.MultiLoc
+getLocation []    = panicPure "Unexpected empty list of tokens"
+getLocation (t:_) = Clang.rangeStart $ Clang.tokenExtent t
+
+addMessage :: DeclId -> DelayedParseMsg -> M ()
+addMessage did msg = modify $ \st -> st {
+      messages = (did, msg) : st.messages
+    }